requirements.txt
streamlit>=1.28.0  # latest works with Py3.13
torch>=2.6.0
torchaudio>=2.6.0
torchvision>=0.15.0
matplotlib>=3.7.0
seaborn>=0.13.0
scikit-learn>=1.2.0
librosa>=0.10.0
numpy>=1.24.0
Pillow>=10.0.0
